refactor: extract PendingTranscriptionBanner.swift
Extracts the PendingTranscriptionBanner View (~53 lines) from
TranscriptRecoveryAndHistoryViews.swift into its own file.

Byte-identical move of lines 3-55. Only imports SwiftUI.

TranscriptRecoveryAndHistoryViews.swift: 130 → 77 lines.
Tests: 667.

⚠️ Manual smoke check recommended: verify pending-transcription banner
renders (needs a session with a queued/failed transcription).
